+++ blocks/property-details-mortgage-calculator/compute-mortgage.js
@@ -1,17 +1,22 @@
-export function computeMortgage(purchasePrice, interestRate = 5, downpaymentPercent = 20, loanTerm = 30) {
-  const downpaymentAmount = downpaymentPercent / 100.0 * purchasePrice;
+export function computeMortgage(
+  purchasePrice,
+  interestRate = 5,
+  downpaymentPercent = 20,
+  loanTerm = 30,
+) {
+  const downpaymentAmount = (downpaymentPercent / 100.0) * purchasePrice;
   const principalAmount = purchasePrice - downpaymentAmount;
-  const interestPayment = interestRate / 100 * principalAmount / 12;
-  var k = interestRate / 100 / 12;
-  var m = Math.pow(1 + k, 12 * loanTerm);
-  var monthlyMortgage = k * m / (m - 1) * principalAmount;
+  const interestPayment = ((interestRate / 100) * principalAmount) / 12;
+  const k = interestRate / 100 / 12;
+  const m = (1 + k) ** (12 * loanTerm);
+  const monthlyMortgage = k * (m / (m - 1)) * principalAmount;
   const principalPayment = monthlyMortgage - interestPayment;
   return {
     downpayment: downpaymentAmount,
     interest: interestPayment,
     principal: principalPayment,
-    totalPayment: monthlyMortgage
-  }
+    totalPayment: monthlyMortgage,
+  };
 }
 
 export function addCommaSeparators(price) {
@@ -19,37 +24,45 @@ export function addCommaSeparators(price) {
 }
 
 export function removeCommas(price) {
-  return price.toString().replace(/,/g,'',);
+  return price.toString().replace(/,/g, '');
 }
 
 export function currencyToNum(price) {
-  return Number(price.toString().replace(/[^0-9.-]+/g,""));
+  return Number(price.toString().replace(/[^0-9.-]+/g, ''));
 }
 
 export function nFormatter(val, a) {
-  var c = [{
+  const c = [
+    {
       value: 1,
-      symbol: ""
-  }, {
+      symbol: '',
+    },
+    {
       value: 1E3,
-      symbol: "k"
-  }, {
+      symbol: 'k',
+    },
+    {
       value: 1E6,
-      symbol: "M"
-  }, {
+      symbol: 'M',
+    },
+    {
       value: 1E9,
-      symbol: "G"
-  }, {
+      symbol: 'G',
+    },
+    {
       value: 1E12,
-      symbol: "T"
-  }, {
+      symbol: 'T',
+    },
+    {
       value: 1E15,
-      symbol: "P"
-  }, {
+      symbol: 'P',
+    },
+    {
       value: 1E18,
-      symbol: "E"
-  }], f;
-  for (f = c.length - 1; 0 < f && !(Number(val) >= c[f].value); f--)
-      ;
-  return (Number(val) / c[f].value).toFixed(a).replace(/\.0+$|(\.[0-9]*[1-9])0+$/, "$1") + c[f].symbol
+      symbol: 'E',
+    },
+  ];
+  let f = c.length - 1;
+  for (; f > 0 && !(Number(val) >= c[f].value); f -= 1);
+  return (Number(val) / c[f].value).toFixed(a).replace(/\.0+$|(\.[0-9]*[1-9])0+$/, '$1') + c[f].symbol;
 }
