+++ blocks/property-details-top/property-details-top.css
@@ -158,222 +158,6 @@
   left: 10px;
 }
 
-.property-image-gallery {
-  overflow: hidden;
-}
-
-.image-gallery-content {
-  position: relative;
-}
-
-.image-gallery-container {
-  max-width: 1600px;
-  margin: 0 auto;
-  overflow: hidden;
-}
-
-.carousel {
-  display: flex;
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  position: relative;
-  left: -75%;
-  transform: translateX(50%);
-}
-
-.carousel.is-reversing {
-  transform: translateX(-50%);
-}
-
-.carousel.is-set {
-  transform: none;
-  transition: transform 0.25s ease 0s;
-}
-
-.carousel-seat {
-  flex: 1 0 50%;
-  order: 2;
-}
-
-.carousel-seat .img-wrapper {
-  padding: 0 8px;
-}
-
-.img-wrapper img {
-  width: 100%;
-  aspect-ratio: 3 / 2;
-  object-fit: cover;
-}
-
-.carousel-seat.is-ref {
-  order: 1;
-}
-
-.nav-row-offset {
-  flex: 0 0 66.667%;
-  max-width: 66.667%;
-  position: relative;
-  min-height: 1px;
-  padding: 0 15px;
-  width: 100%;
-}
-
-.nav-row-icons {
-  justify-content: space-between;
-  display: flex;
-  margin-top: 1rem;
-}
-
-.photos-map {
-  display: flex;
-}
-
-.cmp-map {
-  max-width: 1600px;
-  margin: 0 auto;
-  overflow: hidden;
-  display: flex;
-}
-
-#cmp-map-canvas {
-  flex: 1;
-  width: 100vw;
-  left: 0;
-  margin-bottom: 3.5px;
-}
-
-.invisible {
-  display: none !important;
-}
-
-.button-prop {
-  cursor: pointer;
-  background: var(--white-background);
-  border: 1px solid var(--body-color);
-  line-height: 35px;
-  height: 35px;
-  padding: 0 10px;
-  position: relative;
-  font-family: var(--font-family-proxima);
-  font-weight: 400px;
-  font-size: 12px;
-  text-align: center;
-  white-space: nowrap;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  margin-right: 15px;
-  display: block;
-}
-
-.button-prop.photos {
-  padding-left: 40px;
-}
-
-.button-prop.map {
-  padding-left: 35px;
-}
-
-.button-toggle.disabled {
-  opacity: .5;
-  pointer-events: none;
-}
-
-.icon-photos svg {
-  width: 19px;
-  height: 16px;
-  position: absolute;
-  left: 10px;
-  top: calc(50% - 9px);
-  display: block;
-  pointer-events: none;
-}
-
-.icon-map svg {
-  width: 15px;
-  height: 22px;
-  position: absolute;
-  left: 10px;
-  top: calc(50% - 11px);
-  display: block;
-  pointer-events: none;
-}
-
-.pagination-row {
-  font-size: 14px;
-  text-align: center;
-  line-height: 1;
-  text-transform: uppercase;
-  align-items: center;
-  display: flex;
-  padding-left: 0;
-}
-
-.num-image {
-  margin-right: 0.5rem;
-}
-
-.num-index {
-  font-family: inherit;
-  font-size: inherit;
-  font-weight: inherit;
-  letter-spacing: inherit;
-  line-height: inherit;
-}
-
-button.toggle {
-  margin-right: 0.5rem;
-  cursor: pointer;
-  background-color: var(--white);
-  border: none;
-  outline: none;
-}
-
-button.toggle:focus, button.toggle:active {
-  border: none;
-  outline: none;
-}
-
-.controls svg {
-  height: 19px;
-  width: 11px;
-  outline: none;
-}
-
-.first-carrot svg {
-  transform: scaleX(-1);
-}
-
-.col-12 {
-  flex: 0 0 100%;
-  max-width: 100%;
-}
-
-.col {
-  position: relative;
-  width: 100%;
-  min-height: 1px;
-  padding-right: 15px;
-  padding-left: 15px;
-}
-
-@media (min-width: 576px) {
-  .property-container, .image-gallery-nav {
-    max-width: 540px;
-  }
-}
-
-@media (min-width: 600px) {
-  .property-container, .image-gallery-nav {
-    max-width: 720px;
-  }
-
-  .property-details-info {
-    flex: 0 0 50%;
-    max-width: 50%;
-  }
-}
-
 @media (min-width: 992px) {
   .cmp-property-details-main-attributes-summary {
     padding: 20px 0 40px;
@@ -425,37 +209,5 @@ button.toggle:focus, button.toggle:active {
   .cmp-property-details-main-attributes-summary .btn-save-property, .cmp-property-details-main-attributes-summary .btn-share-property {
     line-height: 40px;
     height: 40px;
-  }
-  
-  .image-gallery-nav {
-    max-width: 960px;
-  }
-
-  .nav-row-offset {
-    margin-left: 8.333%;
-  }
-
-  .pagination-row {
-    font-size: 12px;
-  }
-
-  .button-prop {
-    height: 40px;
-    line-height: 40px;
-  }
-
-  .offset-lg-1 {
-    margin-left: 8.333%
-  }
-
-  .col-lg-10 {
-    flex: 0 0 83.333%;
-    max-width: 83.333%;
-  }
-}
-
-@media (min-width: 1200px) {
-  .property-container, .image-gallery-nav {
-    max-width: 1400px;
   }
 }
