--- blocks/property-details/api.js
@@ -1,11 +1,0 @@
-const propertyAPI = 'https://www.commonmoves.com/bin/bhhs/CregPropertySearchServlet?ucsid=false&SearchType=Radius&ApplicationType=FOR_SALE&Sort=PRICE_ASCENDING&PageSize=9&MinPrice=7497500&MaxPrice=22492500&Latitude=42.56574249267578&Longitude=-70.76632690429688&Distance=2&CoverageZipcode=&teamNearBy=&teamCode=';
-
-export async function getPropertyListing(propID) {
-  const resp = await fetch(propertyAPI);
-  if (resp.ok) {
-    const data = await resp.json();
-    const listing = data.properties.find((item) => item.PropId == propID);
-    return listing;
-  } 
-  return null;
-}
