--- blocks/property-details/property-details-new.css
@@ -1,213 +1,0 @@
-.cmp-property-details-main-attributes-summary {
-  padding: 15px 0 20px;
-}
-
-.property-container .property-container {
-  width: 100% !important;
-  padding-left: 0;
-  padding-right: 0;
-}
-
-.cmp-property-details-main-attributes-summary .hidden {
-  display: none !important;
-}
-
-.cmp-property-details-main-attributes-summary .address {
-  font-family: var(--font-family-georgia);
-  font-weight: 700;
-  font-size: var(--heading-font-size-s);
-  letter-spacing: 0;
-  line-height: 24px;
-}
-
-.cmp-property-details-main-attributes-summary .address h1 {
-  font-size: var(--heading-font-size-s);
-  line-height: 25px;
-  font-family: var(--font-family-georgia);
-  margin: 0;
-}
-
-.cmp-property-details-main-attributes-summary .specs {
-  opacity: .9;
-  font-family: var(--font-family-proxima);
-  font-weight: 400;
-  font-size: 13px;
-  letter-spacing: 0;
-  line-height: 18px;
-}
-
-.cmp-property-details-main-attributes-summary .courtesyOf {
-  font-family: var(--font-family-proxima);
-  font-weight: 300;
-  opacity: 1;
-  letter-spacing: 0;
-  text-align: left;
-  font-size: 15px;
-}
-
-.cmp-property-details-main-attributes-summary .open-house {
-  font-size: var(--body-font-size-xs);
-  color: #457345;
-  letter-spacing: 0;
-  line-height: 30px;
-  flex-direction: column;
-}
-
-.cmp-property-details-main-attributes-summary .open-house .col-auto {
-  padding-right: 0;
-}
-
-.cmp-property-details-main-attributes-summary .open-house__icon {
-  border: 1px solid #457345;
-  color: #457345;
-  display: inline-block;
-  font-family: var(--font-family-proxima);
-  font-weight: 600;
-  padding: 0 8px 0 32px;
-  position: relative;
-}
-
-.cmp-property-details-main-attributes-summary .open-house__icon svg {
-  height: 24px;
-  left: 4px;
-  position: absolute;
-  top: 2px;
-  width: 24px;
-}
-
-.cmp-property-details-main-attributes-summary .open-house__icon use {
-  color: #457345;
-}
-
-.cmp-property-details-main-attributes-summary .open-house__icon .label {
-  text-transform: capitalize;
-  font-family: inherit;
-  font-size: inherit;
-  font-weight: inherit;
-  letter-spacing: inherit;
-  line-height: inherit;
-  color: inherit;
-}
-
-.cmp-property-details-main-attributes-summary .open-house .col-auto {
-  padding-right: 0;
-}
-
-.cmp-property-details-main-attributes-summary .open-house .col-auto:nth-child(2) {
-  padding-left: 8px;
-}
-
-.cmp-property-details-main-attributes-summary .open-house__meta {
-  display: flex;
-  line-height: 15px;
-  flex-direction: column;
-  margin-left: 6px;
-}
-
-.cmp-property-details-main-attributes-summary .open-house__datetime {
-  display: flex;
-  margin-top: 8px;
-}
-
-.cmp-property-details-main-attributes-summary .open-house__meta .date {
-  font-family: var(--font-family-proxima);
-  font-weight: 600;
-  color: #457345;
-}
-
-.cmp-property-details-main-attributes-summary .open-house__meta .time {
-  font-family: var(--font-family-proxima);
-  font-weight: 600;
-  padding-left: 8px;
-  color: #457345;
-}
-
-.cmp-property-details-main-attributes-summary .price {
-  font-family: var(--font-family-georgia);
-  font-weight: 600;
-  font-size: 18px;
-  letter-spacing: -0.35px;
-  line-height: 31px;
-}
-
-.cmp-property-details-main-attributes-summary .btn-save-property, .cmp-property-details-main-attributes-summary .btn-share-property {
-  background: var(--white-background);
-  cursor: pointer;
-  border: 1px solid #aaa;
-  height: 35px;
-  line-height: 35px;
-  padding: 0 15px;
-  position: relative;
-  font-family: var(--font-family-proxima);
-  font-weight: 400;
-  font-size: 12px;
-  padding-left: 45px;
-  text-align: left;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  white-space: normal;
-  border-radius: 0;
-}
-
-.cmp-property-details-main-attributes-summary .btn-save-property svg, .cmp-property-details-main-attributes-summary .btn-share-property svg {
-  display: block;
-  width: 23px;
-  height: 23px;
-  top: calc(50% - 11px);
-  position: absolute;
-  left: 10px;
-}
-
-@media (min-width: 992px) {
-  .cmp-property-details-main-attributes-summary {
-    padding: 20px 0 40px;
-  }
-
-  .cmp-property-details-main-attributes-summary .address {
-    font-family: var(--font-family-georgia);
-    font-weight: 600;
-    font-size: 23px;
-    line-height: 30px;
-  }
-
-  .cmp-property-details-main-attributes-summary .specs {
-    opacity: 1;
-    font-family: var(--font-family-proxima);
-    font-weight: 300;
-    font-size: 15px;
-    line-height: 25px;
-  }
-
-  .cmp-property-details-main-attributes-summary .open-house {
-    flex-direction: row;
-  }
-
-  .cmp-property-details-main-attributes-summary .open-house__meta {
-    line-height: 32px;
-    flex-direction: row;
-    margin-left: 0;
-  }
-
-  .cmp-property-details-main-attributes-summary .open-house__datetime {
-    margin-top: 0;
-  }
-
-  .cmp-property-details-main-attributes-summary .open-house__meta .date, .cmp-property-details-main-attributes-summary .open-house__meta .time {
-    display: inline;
-  }
-
-  details-main-attributes-summary .open-house__datetime .time {
-    margin-right: 8px;
-  }
-
-  .cmp-property-details-main-attributes-summary .price {
-    font-size: 25px;
-    letter-spacing: -0.5px;
-    line-height: 32px;
-  }
-
-  .cmp-property-details-main-attributes-summary .btn-save-property, .cmp-property-details-main-attributes-summary .btn-share-property {
-    line-height: 40px;
-    height: 40px;
-  }
-}
